updated home template

+++ src/main/java/com/btc/web/service/SchedulerService.java
@@ -26,6 +26,9 @@ public class SchedulerService {
     @Value("${playlist.widget}")
     private String playlistWidget;
 
+    @Value("${email.enable}")
+    private boolean emailEnable;
+
     @Autowired
     private PlaylistService playlistService;
 
@@ -40,30 +43,31 @@ public class SchedulerService {
     //@Scheduled(cron = "0 0 0 1/1 * ? *") //Midnight daily
     @Scheduled(cron = "0/10 * * * * ?")
     public void updateWidget() throws Exception {
+        if (emailEnable) {
 
-        WidgetFeedReturn widgetFeedReturn = playlistService.getWidgetFeedInfo();
-        if (widgetFeedReturn != null
-                && playlistService.insertPlaylistItem(playlistWidget, widgetFeedReturn.getVideoId())
-                && playlistService.deletePlaylistItem(widgetFeedReturn.getId())) {
-            //ws.updatePlaylists();
-            widgetFeedReturn.setSuccess(true);
-            widgetFeedReturn.setMessage(new StringBuilder()
-                    .append("Widget updated with '")
-                    .append(widgetFeedReturn.getVideoTitle())
-                    .append("', '")
-                    .append(widgetFeedReturn.getVideoJoke())
-                    .append("'<br>")
-                    .append("There are ")
-                    .append(widgetFeedReturn.getPlaylistCount() - 1)
-                    .append(" videos left to draw from.").toString());
-            logger.info("Widget updated");
-        } else {
-            widgetFeedReturn.setSuccess(false);
-            widgetFeedReturn.setMessage("WIDGET WAS NOT UPDATED, PLEASE CHECK YOUR FEED");
-            logger.error("Widget was not updated");
+            WidgetFeedReturn widgetFeedReturn = playlistService.getWidgetFeedInfo();
+            if (widgetFeedReturn != null
+                    && playlistService.insertPlaylistItem(playlistWidget, widgetFeedReturn.getVideoId())
+                    && playlistService.deletePlaylistItem(widgetFeedReturn.getId())) {
+                //ws.updatePlaylists();
+                widgetFeedReturn.setSuccess(true);
+                widgetFeedReturn.setMessage(new StringBuilder()
+                        .append("Widget updated with '")
+                        .append(widgetFeedReturn.getVideoTitle())
+                        .append("', '")
+                        .append(widgetFeedReturn.getVideoJoke())
+                        .append("'<br>")
+                        .append("There are ")
+                        .append(widgetFeedReturn.getPlaylistCount() - 1)
+                        .append(" videos left to draw from.").toString());
+                logger.info("Widget updated");
+            } else {
+                widgetFeedReturn.setSuccess(false);
+                widgetFeedReturn.setMessage("WIDGET WAS NOT UPDATED, PLEASE CHECK YOUR FEED");
+                logger.error("Widget was not updated");
+            }
+            sendWidgetUpdateAlert(widgetFeedReturn);
         }
-        sendWidgetUpdateAlert(widgetFeedReturn);
-
     }
 
     private void sendWidgetUpdateAlert(WidgetFeedReturn widgetFeedReturn) throws Exception {
